requirements.txt
asgiref==3.7.2
asttokens==2.4.1
boto3==1.34.26
botocore==1.34.26
certifi==2023.11.17
chardet==3.0.4
colorama==0.4.6
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
dnspython==2.5.0
executing==2.0.1
idna==2.10
jedi==0.19.1
jmespath==1.0.1
matplotlib-inline==0.1.6
parso==0.8.3
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.24.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.4
wcwidth==0.2.13
whitenoise==6.6.0
